Improve more on code quality
