+++ duke.txt
@@ -1,3 +1,3 @@
-T | 1 | read books
-D | 1 | asdfasfd | 1212-12-12T12:12
+T | 0 | hii
 T | 0 | read books
+T | 0 | bye
